apps/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
mypy==1.17.0
mypy_extensions==1.1.0
openai==1.97.1
packaging==25.0
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdfium2==4.30.0
python-decouple==3.8
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
